hello , it's parsing and output of the matrix

+++ ft_printf/libft/ft_atoi.c
@@ -6,15 +6,15 @@
 /*   By: user <user@student.42.fr>                  +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2019/04/08 20:12:17 by dbendu            #+#    #+#             */
-/*   Updated: 2019/10/12 17:07:12 by ymanilow         ###   ########.fr       */
+/*   Updated: 2019/10/17 18:37:44 by ymanilow         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "libft.h"
 
-long int	ft_atoi(const char *str)
+__int128_t	ft_atoi(const char *str)
 {
-	long long int	res;
+	__int128_t	res;
 	short int		sign;
 
 	while (ft_isspace(*str))
